chatbox/requirements.txt
python-dotenv
openai
streamlit
google-generativeai
anthropic
cohere
replicate